+++ pytest/test_bem_problems_and_results.py
@@ -20,7 +20,6 @@ from capytaine.bodies.predefined.cylinders import HorizontalCylinder
 
 from capytaine.bem.problems_and_results import LinearPotentialFlowProblem, DiffractionProblem, RadiationProblem, \
     LinearPotentialFlowResult, DiffractionResult, RadiationResult
-from capytaine.bem.solver import Nemoh
 from capytaine.io.xarray import problems_from_dataset, assemble_dataset
 
 from capytaine.io.legacy import import_cal_file
@@ -308,4 +307,12 @@ def test_fill_dataset_with_wavenumbers():
     test_matrix = xr.Dataset(coords={'wavenumber': k_range, 'wave_direction': [0, np.pi/2], 'radiating_dof': ['Heave']})
     dataset = solver.fill_dataset(test_matrix, [body])
     assert np.allclose(dataset.coords['wavenumber'], k_range)
-    print(dataset)
+
+
+def test_fill_dataset_with_periods():
+    body = cpt.FloatingBody(mesh=cpt.mesh_horizontal_cylinder(radius=1, center=(0, 0, -2)),
+                            dofs=cpt.rigid_body_dofs(rotation_center=(0, 0, -2)))
+    T_range = np.linspace(1.0, 3.0, 3)
+    test_matrix = xr.Dataset(coords={'period': T_range, 'wave_direction': [0, np.pi/2], 'radiating_dof': ['Heave']})
+    dataset = solver.fill_dataset(test_matrix, [body])
+    np.testing.assert_allclose(sorted(dataset.coords['period']), sorted(T_range))